requirements.txt
pytest==8.0.0
aider-chat==0.18.0
PyYAML==6.0.1 